requirements.txt
appdirs==1.4.0
arrow==0.5.0
beautifulsoup4==4.3.2
bibtexparser==0.6.0
diff-match-patch==20121119
dj-database-url==0.3.0
Django==1.10.2
django-admin-tools==0.8.0
django-appconf==1.0.2
django-compressor==2.1
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
django-extensions==1.5.5
django-filter==0.15.3
django-filters==0.2.1
django-import-export==0.5.1
django-static-precompiler==1.5
django-staticfiles==1.2.1
django-tastypie==0.13.3
djangorestframework==3.5.1
djangorestframework-filters==0.9.1
djangorestframework-queryfields==1.0.0
dropbox==2.2.0
flake8==3.0.4
flake8-docstrings==1.0.2
future==0.14.3
lxml==3.7.3
Markdown==2.6.7
mccabe==0.5.2
memoized-property==1.0.2
mendeley==0.3.2
oauthlib==0.7.2
packaging==16.8
Pillow==2.6.2
psycopg2==2.6
pycodestyle==2.0.0
pydocstyle==1.1.1
pyflakes==1.2.3
pyparsing==2.1.10
python-dateutil==2.5.3
python-decouple==2.3
python-docx==0.8.5
python-mimeparse==1.6.0
pytz==2015.4
rcssmin==1.0.6
requests==2.7.0
requests-oauthlib==0.4.2
rjsmin==1.0.12
six==1.10.0
tablib==0.11.4
Unipath==1.1
urllib3==1.20
xlwt==1.1.2
